+++ instrumentation/spring/spring-batch-3.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/spring/batch/step/StepExecutionTracer.java
@@ -26,10 +26,6 @@ public class StepExecutionTracer extends BaseTracer {
     return Context.current().with(span);
   }
 
-  public void end(Context context) {
-    end(Span.fromContext(context));
-  }
-
   @Override
   protected String getInstrumentationName() {
     return "io.opentelemetry.javaagent.spring-batch";
